refactor(*): 패키지 구조 변경 및 인텐트로 데이터로부터 받지 못한 경우 onCreate 딴에서 바로 finish하기
