+++ app/src/main/AndroidManifest.xml
@@ -12,16 +12,16 @@
         android:theme="@style/Theme.Movie"
         tools:targetApi="31">
         <activity
-            android:name=".activity.ScreeningSeatSelectionActivity"
+            android:name=".seatselection.ScreeningSeatSelectionActivity"
             android:exported="false"/>
         <activity
-            android:name=".activity.ReservationResultActivity"
+            android:name=".reservationresult.ReservationResultActivity"
             android:exported="false" />
         <activity
-            android:name=".activity.reservation.ReservationActivity"
+            android:name=".reservation.ReservationActivity"
             android:exported="false" />
         <activity
-            android:name=".activity.movies.MoviesActivity"
+            android:name=".movies.MoviesActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
