+++ app/src/main/java/woowacourse/movie/util/IntentUtil.kt
@@ -5,7 +5,7 @@ import android.os.Build
 import java.io.Serializable
 
 @Suppress("DEPRECATION")
-inline fun <reified T : Serializable> Intent.customGetSerializable(key: String): T? {
+inline fun <reified T : Serializable> Intent.getSerializableExtraByKey(key: String): T? {
     return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
         getSerializableExtra(key, T::class.java)
     } else {
