+++ app/src/main/res/layout/activity_reservation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activity.reservation.ReservationActivity">
+    tools:context=".reservation.ReservationActivity">
 
     <ScrollView
         android:id="@+id/movie_information_view"
