+++ app/src/main/res/layout/activity_reservation_result.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activity.ReservationResultActivity">
+    tools:context=".reservationresult.ReservationResultActivity">
 
     <TextView
         android:layout_width="match_parent"
